+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dysporium-sdk/provider
 
+## 2.0.0
+
+### Major Changes
+
+- update for more cases
+
 ## 1.0.0
 
 ### Major Changes