+++ packages/qwen/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @dysporium-sdk/qwen
 
+## 2.0.0
+
+### Major Changes
+
+- update for more cases
+
+### Patch Changes
+
+- Updated dependencies
+  - @dysporium-sdk/core@2.0.0
+  - @dysporium-sdk/provider@2.0.0
+
 ## 1.1.1
 
 ### Patch Changes